update all spark 3
