+++ hudi-spark-datasource/hudi-spark3.0.x/src/main/scala/org/apache/spark/sql/HoodieSpark30CatalystPlanUtils.scala
@@ -50,8 +50,12 @@ object HoodieSpark30CatalystPlanUtils extends HoodieSpark3CatalystPlanUtils {
     plan match {
       case s@ScanOperation(_, _,
       l@LogicalRelation(fs: HadoopFsRelation, _, _, _)) if fs.fileFormat.isInstanceOf[NewHoodieParquetFileFormat] && !fs.fileFormat.asInstanceOf[NewHoodieParquetFileFormat].isProjected =>
-        fs.fileFormat.asInstanceOf[NewHoodieParquetFileFormat].isProjected = true
-        Project(l.resolve(fs.location.asInstanceOf[SparkHoodieTableFileIndex].schema, fs.sparkSession.sessionState.analyzer.resolver), s)
+        val ff = fs.fileFormat.asInstanceOf[NewHoodieParquetFileFormat]
+        ff.isProjected = true
+        val tableSchema = fs.location.asInstanceOf[SparkHoodieTableFileIndex].schema
+        val resolvedSchema = l.resolve(tableSchema, fs.sparkSession.sessionState.analyzer.resolver)
+        val projection: LogicalPlan = Project(resolvedSchema, s)
+        HoodieSpark3CatalystPlanUtils.applyFiltersToPlan(projection, tableSchema, resolvedSchema, ff.getRequiredFilters)
       case _ => plan
     }
   }
